Update create customer
