+++ practice-two/src/components/FormValidation/FormValidation.jsx
@@ -1,18 +1,112 @@
-import { memo, useContext } from 'react';
+import { memo, useState, useContext, useCallback } from 'react';
 import './FormValidation.css';
 import { InputValidate, Button } from '@components';
 import { BUTTON_VARIANTS } from '@constants';
-import { FormContext } from '@contexts';
+import { Validation } from '@helpers';
+import { createCustomerService } from '@services';
+import { CustomerContext } from '@contexts';
 
-const FormValidation = () => {
-  const {
-    formData,
-    errors,
-    handleShowForm,
-    handleChange,
-    handleBlur,
-    handleSubmit,
-  } = useContext(FormContext);
+const FormValidation = ({
+  handleShowForm,
+  setCreateResult,
+}) => {
+  const [formData, setFormData] = useState({
+    name: '',
+    avatar: '',
+    mail: '',
+    phoneNumber: '',
+    description: '',
+    address: '',
+    gender: 'Male',
+  });
+  const [errors, setErrors] = useState({
+    name: '',
+    avatar: '',
+    mail: '',
+    phoneNumber: '',
+    description: '',
+    address: '',
+  });
+  const { handleCreateCustomer } = useContext(CustomerContext);
+
+  // Set value for form data
+  const handleChange = useCallback(
+    (e) => {
+      const { name, value } = e.target;
+      setFormData((prevFormData) => ({
+        ...prevFormData,
+        [name]: value,
+      }));
+    },
+    [setFormData]
+  );
+
+  // Check validate input when click outside
+  const handleBlur = useCallback(
+    (e) => {
+      const { name, value } = e.target;
+      const errorMessage = Validation(name, value);
+      setErrors((prevErrors) => ({
+        ...prevErrors,
+        [name]: errorMessage,
+      }));
+    },
+    [setErrors]
+  );
+
+  const resetForm = () => {
+    setFormData({
+      name: '',
+      avatar: '',
+      mail: '',
+      phoneNumber: '',
+      description: '',
+      address: '',
+      gender: 'Male',
+    });
+    setErrors({
+      name: '',
+      avatar: '',
+      mail: '',
+      phoneNumber: '',
+      description: '',
+      address: '',
+    });
+  };
+
+  // Submit form
+  const handleSubmit = useCallback(
+    async (e) => {
+      e.preventDefault();
+
+      // Check validation form when submiting
+      const newErrors = {};
+      for (const field in formData) {
+        const errorMessage = Validation(field, formData[field]);
+        if (errorMessage) {
+          newErrors[field] = errorMessage;
+        }
+      }
+      setErrors(newErrors);
+
+      if (Object.keys(newErrors).length === 0) {
+        const response = await createCustomerService(formData);
+
+        if (response.error) {
+          handleShowForm();
+          setCreateResult('failed');
+          resetForm();
+          return;
+        } else {
+          handleShowForm();
+          handleCreateCustomer(formData)
+          setCreateResult('success');
+          resetForm();
+        }
+      }
+    },
+    [formData, setErrors, handleShowForm]
+  );
 
   return (
     <div className='form__background'>
