+++ practice-two/src/contexts/index.js
@@ -1,2 +1,1 @@
 export { CustomerProvider, CustomerContext } from './CustomerContext';
-export { FormProvider, FormContext } from './FormContext';
