+++ practice-two/src/main.jsx
@@ -2,14 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './styles/index.css';
-import { CustomerProvider, FormProvider } from '@contexts';
+import { CustomerProvider } from '@contexts';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <CustomerProvider>
-      <FormProvider>
-        <App />
-      </FormProvider>
+      <App />
     </CustomerProvider>
   </React.StrictMode>
 );
